Refactor collector
